better code reuse for data types
