+++ revengai/features/__init__.py
@@ -188,187 +188,6 @@ class BaseDialog(QDialog):
                 ),
             )
 
-    # def _function_import_symbol_datatypes(
-    #         self,
-    #         matched_function_bid: int = 0,
-    #         matched_func_id: int = 0,
-    #         function_addr: int = 0
-    # ) -> None:
-    #     def apply_type(deci: DecompilerInterface, artifact) -> None | str:
-    #         supported_types = [
-    #               Function,
-    #               GlobalVariable,
-    #               Enum,
-    #               Struct,
-    #               Typedef
-    #         ]
-
-    #         if not any(isinstance(artifact, t) for t in supported_types):
-    #             return "Unsupported artifact type: " \
-    #                    f"{artifact.__class__.__name__}"
-
-    #         if isinstance(artifact, Function):
-    #             deci.functions[artifact.addr] = artifact
-    #         elif isinstance(artifact, GlobalVariable):
-    #             deci.global_vars[artifact.addr] = artifact
-    #         elif isinstance(artifact, Enum):
-    #             deci.enums[artifact.name] = artifact
-    #         elif isinstance(artifact, Struct):
-    #             deci.structs[artifact.name] = artifact
-    #         elif isinstance(artifact, Typedef):
-    #             deci.typedefs[artifact.name] = artifact
-
-    #         return None
-
-    #     def apply_types(
-    #             deci: DecompilerInterface,
-    #             artifacts: list
-    #     ) -> None | str:
-    #         for artifact in artifacts:
-    #             error = apply_type(deci, artifact)
-    #             if error is not None:
-    #                 return error
-    #         return None
-
-    #     def _load_many_artifacts_from_list(artifacts: list[dict]) -> list:
-    #         _artifacts = []
-    #         for artifact in artifacts:
-    #             art = _art_from_dict(artifact)
-    #             if art is not None:
-    #                 _artifacts.append(art)
-    #         return _artifacts
-
-    #     try:
-    #         # first step is to get the analysis id for the function
-    #         res: dict = RE_analysis_lookup(matched_function_bid).json()
-    #         matched_analysis_id = res.get("analysis_id", 0)
-
-    #         if matched_analysis_id == 0:
-    #             logger.error(
-    #                 "Failed to get analysis id for functionId %d.",
-    #                 matched_func_id,
-    #             )
-    #             return
-
-    #         # second step is to start the generation of the datatypes
-    #         res = RE_generate_data_types(
-    #             matched_analysis_id,
-    #             [matched_func_id]
-    #         ).json()
-    #         status = res.get("status", False)
-
-    #         if status:
-    #             logger.info(
-    #                 "Successfully started the generation of functions"
-    #                 " data types"
-    #             )
-    #         else:
-    #             logger.error(
-    #                 "Failed to start the generation of functions data types"
-    #             )
-    #             return
-
-    #         # try list the datatypes
-
-    #         res: dict = RE_list_data_types(
-    #             matched_analysis_id,
-    #             [matched_func_id]
-    #         ).json()
-
-    #         status = res.get("status", False)
-
-    #         if not status:
-    #             logger.error("Error getting function data types")
-    #             return
-
-    #         # TODO: remove this
-    #         logger.info(f"Let's stop here at the moment... {res}")
-    #         return
-
-    #         deci = DecompilerInterface.discover(force_decompiler="ida")
-    #         if not deci:
-    #             logger.error("Libbs: Unable to find a decompiler")
-    #             return
-
-    #         total_count = res.get("data", {}).get("total_count", 0)
-
-    #         if total_count > 0:
-    #             items = res.get("data", {}).get("items", [])
-    #             for item in items:
-    #                 function_types = item.get(
-    #                     "data_types", {}).get("func_types", None)
-    #                 func_deps = item.get(
-    #                     "data_types", {}).get("func_deps", [])
-    #                 function_id = item.get("function_id", 0)
-
-    #                 if function_types and len(func_deps) > 0:
-
-    #                     deps = _load_many_artifacts_from_list(
-    #                         func_deps,
-    #                     )
-
-    #                     logger.info(
-    #                         f"Loaded {len(func_deps)} for function "
-    #                         f"{function_id}"
-    #                     )
-
-    #                     deps_res = apply_types(deci, deps)
-    #                     if deps_res is not None:
-    #                         logger.error(
-    #                             "Error applying data type dependencies for "
-    #                             f"function {function_id}: {deps_res}"
-    #                         )
-    #                         return
-    #                     else:
-    #                         logger.info(
-    #                             "Applied data type dependencies for function"
-    #                             f" id {function_id}"
-    #                         )
-
-    #                 if function_types:
-    #                     func: Function = _art_from_dict(function_types)
-    #                     # repalce function address with the one we need to
-    #                     # apply the data type to
-    #                     func.addr = function_addr
-    #                     func_res = apply_type(deci, func)
-    #                     if func_res is not None:
-    #                         logger.error(
-    #                             "Error applying function data type for "
-    #                             f"function id {function_id}: {func_res}"
-    #                         )
-    #                         return
-    #                     else:
-    #                         logger.info(
-    #                             "Applied data types for function id "
-    #                             f"{function_id}"
-    #                         )
-
-    #             logger.info("Function data types application completed")
-    #         else:
-    #             logger.warning("No function data types to apply")
-
-    #     except HTTPError as e:
-    #         error = e.response.json().get(
-    #             "error",
-    #             f"An unexpected error occurred. Sorry for the "
-    #             f"inconvenience. {e.response.status_code}",
-    #         )
-
-    #         logger.error(
-    #             "Error while importing data types for functionId "
-    #             f"{matched_func_id}: {error}"
-    #         )
-
-    #         inmain(idaapi.warning, error)
-
-    #     except ValueError as e:
-    #         logger.error(
-    #             "Error while importing data types for functionId "
-    #             f"{matched_func_id}: {e}"
-    #         )
-
-    #         inmain(idaapi.warning, str(e))
-
     def _function_rename(
             self, func_addr: int, new_func_name: str, func_id: int = 0
     ) -> None:
